Start brokers.api module
